+++ AproxymationExtrapolation/Test/databasemock.h
@@ -6,8 +6,8 @@
 class DataBaseMock : public DataBase
 {
 public:
-    MOCK_METHOD1(setDataOfOrdinates, void(const float));
-    MOCK_METHOD1(setDataOfSevered, void(const float));
+    MOCK_METHOD1(setDataOfOrdinates, void(const double));
+    MOCK_METHOD1(setDataOfSevered, void(const double));
     MOCK_CONST_METHOD0(getVectorOfOrdinates, const QVector<double>&());
     MOCK_CONST_METHOD0(getVectorOfSevered, const QVector<double>&());
     MOCK_METHOD0(clearDataBase, void());
